pyproject.toml
[project]
name = "MBA"
version = "0.1.0"
description = "END to END MBA Pipeline"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.28.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "streamlit>=1.49.1",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "sqlalchemy>=2.0.43",
    "pymysql>=1.1.2",
    "tabulate>=0.9.0",
    "click>=8.0.0",
    "git-filter-repo>=2.47.0",
    "strands-agents>=1.10.0",
    "strands-agents-tools>=0.2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
# CLI tools you already had
mba-ingest = "MBA.cli:main"
mba-producer = "MBA.microservices.producer:main"
mba-worker = "MBA.microservices.worker:main"
mba-api = "MBA.microservices.api:run_server"
mba-db = "MBA.database.cli:main"

# NEW: Streamlit launcher (calls MBA.streamlit_app:main)
mba-app = "MBA.app_launcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/MBA"]
